+++ .gitignore
@@ -5,7 +5,6 @@ target/
 
 ### IntelliJ IDEA ###
 .target/
-.idea
 .idea/modules.xml
 .idea/jarRepositories.xml
 .idea/compiler.xml
