+++ pom.xml
@@ -14,9 +14,19 @@
     </properties>
     <dependencies>
         <dependency>
-            <groupId>io.netty</groupId>
-            <artifactId>netty-all</artifactId>
-            <version>4.1.75.Final</version>
+            <groupId>com.ditchoom</groupId>
+            <artifactId>buffer-iosx64</artifactId>
+            <version>1.0.72</version>
+        </dependency>
+        <dependency>
+            <groupId>com.ditchoom</groupId>
+            <artifactId>buffer-tvosx64</artifactId>
+            <version>1.0.72</version>
+        </dependency>
+        <dependency>
+            <groupId>com.ditchoom</groupId>
+            <artifactId>buffer</artifactId>
+            <version>1.0.72</version>
         </dependency>
     </dependencies>
