--- src/ClientDirectory/Garbage/Trash/1.txt
@@ -1,1 +1,0 @@
-sdsdsdsd
