--- src/ServerDirectory/Server.txt
@@ -1,1 +1,0 @@
-wwwwwwwwwwwwwwwwwwwqwqwqwqwqw
