+++ src/main/java/Message.java
@@ -1,0 +1,2 @@
+public abstract class Message {
+}
