--- src/main/java/Storage/CloudServer.java
@@ -1,47 +1,0 @@
-package Storage;
-
-import io.netty.bootstrap.ServerBootstrap;
-import io.netty.channel.ChannelFuture;
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.ChannelOption;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.nio.NioEventLoopGroup;
-import io.netty.channel.socket.nio.NioServerSocketChannel;
-import io.netty.handler.codec.serialization.ClassResolvers;
-import io.netty.handler.codec.serialization.ObjectDecoder;
-import io.netty.handler.codec.serialization.ObjectEncoder;
-
-
-public class CloudServer {
-    public void run() throws Exception{
-        EventLoopGroup mainGroup = new NioEventLoopGroup();
-        EventLoopGroup workerGroup = new NioEventLoopGroup();
-        try {
-            ServerBootstrap b = new ServerBootstrap();
-            b.group(mainGroup, workerGroup)
-                    .channel(NioServerSocketChannel.class)
-                   .childOption(ChannelOption.SO_KEEPALIVE, true)
-                    .childHandler(new ChannelInitializer<io.netty.channel.socket.SocketChannel>() {
-                        @Override
-                        protected void initChannel(io.netty.channel.socket.SocketChannel socketChannel) {
-                            socketChannel.pipeline().addLast(
-                                   new ObjectDecoder(50 * 1024 * 1024, ClassResolvers.cacheDisabled(null)),
-                                    new ObjectEncoder(),
-                                    new ServerHandler());
-                        }
-                    });
-            ChannelFuture future = b.bind(8189).sync();
-            System.out.println("Server started...");
-            future.channel().closeFuture().sync();
-        } finally {
-            mainGroup.shutdownGracefully();
-            workerGroup.shutdownGracefully();
-        }
-    }
-
-    public static void main(String[] args) throws Exception {
-        new CloudServer().run();
-
-    }
-
-}
